feat(UserInfo)[OK-267] Add decoded token to AuthContext
