+++ client/src/app/Providers.tsx
@@ -4,10 +4,11 @@ import React from "react";
 import { AuthProvider } from "@/contexts/AuthContext";
 import { ThemeProvider } from "next-themes";
 import ToasterWrapper from "@/providers/ToasterWrapper";
+import { DecodedToken } from "@/interfaces/decoded-token/decoded-token";
 
-const Providers = ({ children }: { children: React.ReactNode }) => {
+const Providers = ({ children, decodedToken }: { children: React.ReactNode; decodedToken: DecodedToken | null }) => {
   return (
-    <AuthProvider>
+    <AuthProvider decodedToken={decodedToken}>
       <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
         <ToasterWrapper />
         {children}
