+++ client/src/app/user-boards/(realtime-canvas)/[boardId]/slides/[slideIndex]/page.tsx
@@ -8,7 +8,7 @@ import SlideFileInputs from "@/components/board/SlideFileInputs";
 const Board: React.FC = () => {
   return (
     <div className="flex flex-col">
-      <BoardHorizontalMenu boardName={"Board 1"} groupId={MenuGroups.FILE_AND_CANVAS_OPERATIONS} />
+      <BoardHorizontalMenu groupId={MenuGroups.FILE_AND_CANVAS_OPERATIONS} />
       <div className="flex">
         <BoardSidebar withSettings={true} groupId={MenuGroups.DRAWING_TOOLS} />
         <BoardSidebar groupId={MenuGroups.OBJECT_MANIPULATION} />
